Added a functionnal core
I added the foundations of the website.
The foundations are heavily based on the Symfony MVC model.

The Core contains all the importants stuff mostly the routing system that redirect what page is shown.
Then we have the Controllers that are class that will launch their functions when you try to load a page.
Very usefull if you need to do somes specifics operations each time someone call a page.

We also have the views page where the components that allows the creations of pages are.

The config file is where we will store all the sensible data.

Routes is where we declare the valild URL path.

lang will be for an easier and more generalist translations of the pages

and public is also where we will store all the public informations